% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_helpers.R
\name{forecast_helpers}
\alias{forecast_helpers}
\alias{ts_fitter}
\alias{ts_forecaster}
\title{A Collection of Utility Functions}
\usage{
ts_fitter(.tbl, .fun, ...)

ts_forecaster(.split, .model)
}
\arguments{
\item{.tbl}{The \code{tibble} to be analized}

\item{.fun}{The function to be applied}

\item{...}{Additional arguments to be passed to \code{...}.}

\item{.split}{The columns that contains the splits}

\item{.model}{The model to be forecasted}
}
\description{
A Collection of Utility Functions
}
