% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Summarise Forecasting Models Errror Metrics}
\usage{
get_metrics(.tbl, .group, .truth, .forecast)
}
\arguments{
\item{.tbl}{A \code{tibble} of the \code{tbl_time} class.}

\item{.group}{The column in which the data should be grouped. This will often be a column with stock tickers or stocks names.}

\item{.truth}{The truth results of a time series realization.}

\item{.forecast}{The out-of-sample forecasts from the chosen model.}
}
\value{
A \code{tibble} with the following summary statistics by asset: mse, rmse, mae, mape and mase.
}
\description{
Summarise the out-of-sample stocks forecasts by asset and time periods.
}
\examples{
library(YahooTickers)
library(dplyr)
library(forecast)

get_tickers(dow) \%>\%
  slice(2:3) \%>\%
  get_stocks(., periodicity = "monthly") \%>\%
  get_returns(., tickers, log, TRUE, adjusted) \%>\%
  get_models(., tickers, adjusted, 100, 1, FALSE, auto.arima,
             seasonal = FALSE, stationary = TRUE) \%>\%
  get_forecasts(.) \%>\%
  get_metrics(., tickers, adjusted, point_forecast)
}
