% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_models.R
\name{get_models}
\alias{get_models}
\alias{get_models.default}
\alias{get_models.tbl_df}
\alias{get_models.tbl_time}
\title{Extract in-sample statistics estimated in "real-time"}
\usage{
get_models(.tbl, .group, .col, .initial, .assess, .cumulative, .fun, ...)

\method{get_models}{default}(.tbl, .group, .col, .initial, .assess,
  .cumulative, .fun, ...)

\method{get_models}{tbl_df}(.tbl, .group, .col, .initial, .assess,
  .cumulative, .fun, ...)

\method{get_models}{tbl_time}(.tbl, .group, .col, .initial, .assess,
  .cumulative, .fun, ...)
}
\arguments{
\item{.tbl}{A tidy \code{tibble} or an object of the \code{tbl_time} class.}

\item{.group}{The column in which the data should be grouped. This will often be a column with stock tickers or stocks names.}

\item{.col}{The reference column in which the operations should be based on.}

\item{.initial}{The number of periods used to train the model in each split.}

\item{.assess}{The forecast horizon.}

\item{.cumulative}{If \code{TRUE} forecasts accumulate from the origin of each split. If \code{FALSE} forecasts are evaluated in a rolling windown. The default is TRUE.}

\item{.fun}{Currently supports only a forecasting function from the \code{forecast} package.}

\item{...}{Additional parameters to be passed to \code{.fun}.}
}
\value{
A tidy \code{tibble} contaning the following in-sample statistics for each period:

  \itemize{
    \item{term}: The parameters estimated by \code{.fun}
    \item{estimate}: The values of the estimated parameters
    \item{model.desc}: A description of the model including the three integer components (p, d, q) are the AR order, the degree of differencing, and the MA order.
    \item{sigma}: The square root of the estimated residual variance
    \item{logLik}: The data's log-likelihood under the model
    \item{AIC}: The Akaike Information Criterion
    \item{BIC}: The Bayesian Information Criterion
    \item{ME}: Mean error
    \item{RMSE}: Root mean squared error
    \item{MAE}: Mean absolute error
    \item{MPE}: Mean percentage error
    \item{MAPE}: Mean absolute percentage error
    \item{MASE}: Mean absolute scaled error
    \item{ACF1}: Autocorrelation of errors at lag 1
}
}
\description{
\code{get_models()} recursively estimates time-series models by using rolling and expanding time windows.
}
\details{
Under the hood, get_forecasts uses \code{rsample} infrastructure to sequentially access the model passed to \code{.fun}. The data y1, y2, ..., yt is fitted and the process repeats itself for each stock in the \code{.group} column.
}
\examples{
library(YahooTickers)
library(dplyr)
library(forecast)
# Download and forecast time series using the "auto.arima" function
# from the forecast package
get_tickers(dow) \%>\%
  slice(1:3) \%>\%
  get_stocks(., periodicity = "monthly") \%>\%
  get_returns(., tickers, arithmetic, TRUE, adjusted) \%>\%
  get_models(., tickers, return_adjusted, 60, 1, FALSE, auto.arima,
             seasonal = FALSE, stationary = TRUE)
}
\seealso{
\href{https://business-science.github.io/sweep/reference/sw_tidy.html}{sw_tidy()}
         \href{https://business-science.github.io/sweep/reference/tidiers_arima.html}{sw_glance()}
         \href{https://tidymodels.github.io/rsample/reference/rolling_origin.html}{rolling_origin()}
}
