% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tickers.R
\name{get_tickers}
\alias{get_tickers}
\title{Get The Tickers Required by the Yahoo API}
\usage{
get_tickers(exchange)
}
\arguments{
\item{exchange}{A desired exchange. Dee Details.}
}
\value{
A \code{tibble} with stocks' codes used in Yahoo Finance website.
}
\description{
\code{get_tickers()} helps you to get the tickers of a specific stock index.
}
\details{
Currently \code{get_tickers} supports the following stock indexes:

\itemize{
  \item \code{dow}: Dow Jones Index (United States)
  \item \code{sp500}: S&P 500 Index (United States)
  \item \code{nasdaq}: Nasdaq Composite Index (United States)
  \item \code{nyse}: NYSE Composite Index (United States)
  \item \code{amex}: AMEX Composite Index (United States)
  \item \code{russell2000}: Russell 2000 Index (United States)
  \item \code{ftse100}: FTSE 100 Index (United Kingdom)
  \item \code{dax}: Deutsche Boerse AG Index (Germany)
  \item \code{cac40}: CAC 40 Index (France)
  \item \code{bel20}: BEL 20 (Belgium)
  \item \code{topix}: Topix Index (Tokyo)
  \item \code{hangseng}: Hang Seng Index (Hong Kong)
  \item \code{sensex}: BSE Sensex Index (India)
  \item \code{jakarta}: Jakarta Stock Exchange Composite Index (Indonesia)
  \item \code{bursa}: FTSE Bursa Malaysia KLCI Index (Malaysya)
  \item \code{nzx50}: NZX 50 Index (New Zealand)
  \item \code{kospi}: KOSPI Composite Index (South Korea)
  \item \code{taiex}: TWSE Taiwan
  \item \code{tsx}: Toronto Composite Index (Canada)
  \item \code{ibovespa}: Ibovespa Index (Brazil)
  \item \code{ipc}: BMV IPC Index (Mexico)
  \item \code{ipsa}: IPSA Index (Chile)
  \item \code{merval}: Merval Index (Argentina)
}
}
\examples{
library(YahooTickers)

# get the brazilian stocks index tickers
get_tickers(ibovespa)

# get the sp500 stocks tickers
get_tickers(sp500)

# The exchange argument should be unquoted. This throws an error.
# get_tickers("sp500")
}
