% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_helpers.R
\name{getters}
\alias{getters}
\alias{get_index_col}
\alias{get_index_char}
\alias{get_index_quo}
\title{Getters}
\usage{
get_index_col(.tbl)

get_index_char(.tbl)

get_index_quo(.tbl, .index = get_index_char(.tbl))
}
\arguments{
\item{.tbl}{A `YahooTickers` object.}

\item{.index}{The date column.}
}
\description{
Accessors to attributes of `YahooTickers` objects.
}
