% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{get_sp500}
\alias{get_dow}
\alias{get_nasdaq}
\alias{get_nyse}
\alias{get_amex}
\alias{get_russell2000}
\alias{get_ftse100}
\alias{get_dax}
\alias{get_cac40}
\alias{get_bel20}
\alias{get_topix}
\alias{get_hangseng}
\alias{get_sensex}
\alias{get_jakarta}
\alias{get_bursa}
\alias{get_nzx50}
\alias{get_kospi}
\alias{get_taiex}
\alias{get_tsx}
\alias{get_ibov}
\alias{get_ipc}
\alias{get_ipsa}
\alias{get_merval}
\title{A Collection of Utility Functions}
\usage{
get_sp500()

get_dow()

get_nasdaq()

get_nyse()

get_amex()

get_russell2000()

get_ftse100()

get_dax()

get_cac40()

get_bel20()

get_topix()

get_hangseng()

get_sensex()

get_jakarta()

get_bursa()

get_nzx50()

get_kospi()

get_taiex()

get_tsx()

get_ibov()

get_ipc()

get_ipsa()

get_merval()
}
\description{
A Collection of Utility Functions
}
