% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecasts.R
\name{get_forecasts}
\alias{get_forecasts}
\alias{get_forecasts.default}
\alias{get_forecasts.tbl}
\alias{get_forecast.tbl_time}
\title{True Out-of-Sample Time Series Forecasting}
\usage{
get_forecasts(.tbl, ...)

\method{get_forecasts}{default}(.tbl, ...)

\method{get_forecasts}{tbl}(.tbl, ...)

get_forecast.tbl_time(.tbl, ...)
}
\arguments{
\item{.tbl}{A tidy \code{tibble} of the \code{tbl_time} class.}

\item{...}{Additional parameters to be passed to \code{forecast}.}
}
\value{
A tidy \code{tibble}.
}
\description{
\code{get_forecasts()} extracts the true out-of-sample forecasts from an object previously fitted with \code{get_models()}.
}
\examples{
library(YahooTickers)
library(dplyr)
library(forecast)

# Download and forecast time series using the "auto.arima"
# function from the forecast package
get_tickers(dow) \%>\%
  slice(1:2) \%>\%
  get_stocks(., periodicity = "monthly") \%>\%
  get_returns(., tickers, log, TRUE, adjusted) \%>\%
  get_models(., tickers, adjusted, 100, 1, FALSE, auto.arima,
             seasonal = FALSE, stationary = TRUE) \%>\%
  get_forecasts(.)
}
\seealso{
\href{https://tidymodels.github.io/rsample/reference/rolling_origin.html}{rolling_origin()}
}
