% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returns.R
\name{get_returns}
\alias{get_returns}
\title{Calculate Returns From Stock Prices}
\usage{
get_returns(.tbl, ...)
}
\arguments{
\item{.tbl}{A \code{tibble} with stock prices.}

\item{...}{The columns in which the calculation shoud be conducted.}
}
\value{
A tidy \code{tibble}.
}
\description{
Stock prices are known for not been stationary. For econometric modeling use \code{get_returns()}.
}
\examples{
library(YahooTickers)
library(dplyr)
get_tickers("dow") \%>\%
  get_stocks(.) \%>\%
  get_returns(., adjusted)
}
