% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returns.R
\name{get_returns}
\alias{get_returns}
\title{Calculate Returns From Stock Prices}
\usage{
get_returns(.tbl, .group, .type, ...)
}
\arguments{
\item{.tbl}{A tidy \code{tibble}.}

\item{.group}{The column in which the data should be grouped. This will often be a column containing stock tickers or stocks names.}

\item{.type}{The method used to calculate returns. One of: \code{log} or \code{arithmetic}.}

\item{...}{The column (columns) in which the calculation shoud be conducted.}
}
\value{
A tidy \code{tibble}.
}
\description{
Stock prices are known for been non-stationary. For econometric modeling use \code{get_returns()}.
}
\examples{
library(YahooTickers)
library(dplyr)
get_tickers(dow) \%>\%
  get_stocks(.) \%>\%
  get_returns(., tickers, arithmetic, adjusted)
}
