% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stocks.R
\name{get_stocks}
\alias{get_stocks}
\title{A wrapper over \code{getSymbols} that maps over index components.}
\usage{
get_stocks(
  tickers,
  from = "2010-01-01",
  to = Sys.Date(),
  periodicity = "daily",
  simplify = TRUE,
  otherwise = NULL,
  quiet = TRUE
)
}
\arguments{
\item{tickers}{A \code{character} vector with the stock tickers.}

\item{from}{A date in the format \code{YYYY-MM-DD}.}

\item{to}{A date in the format \code{YYYY-MM-DD}.}

\item{periodicity}{One of: \code{daily}, \code{weely} or \code{monthly}.}

\item{simplify}{If TRUE, a \code{tibble} is returned. If FALSE a list with two elements is returned:

  \itemize{
    \item \code{result}: with the downloads that succeeded;
    \item \code{error}: with the stocks not covered by the Yahoo Finance.
  }

The defaul is TRUE.}

\item{otherwise}{Argument passed to \code{\link[purrr]{safely}}.}

\item{quiet}{Argument passed to \code{\link[purrr]{safely}}. A notification is showed whenever an error occurs. Default is \code{TRUE}.}
}
\value{
A tidy \code{tibble} if \code{simplify = TRUE} and a \code{list} of two components if \code{simplify = FALSE}.
}
\description{
\code{get_stocks()} allows you to quickly download stock indexes from the the \href{https://finance.yahoo.com/}{Yahoo Finance} without having to run \code{for} loops.
}
\examples{
# Get stocks from the Merval Stock Index (Argentina)
library(YahooTickers)
ticks <- get_tickers(merval)
get_stocks(ticks, periodicity = "monthly")
}
\seealso{
\href{https://purrr.tidyverse.org/reference/safely.html}{safely()}
}
